APIs/azure.com/apimanagement-apimcertificates/2016-10-10/python-aiohttp-client/tests/test_certificates_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.certificate_collection import CertificateCollection
from openapi_server.models.certificate_contract import CertificateContract
from openapi_server.models.certificate_create_or_update_parameters import CertificateCreateOrUpdateParameters
from openapi_server.models.certificates_list_by_service_default_response import CertificatesListByServiceDefaultResponse


pytestmark = pytest.mark.asyncio

async def test_certificates_create_or_update(client):
    """Test case for certificates_create_or_update

    
    """
    parameters = {"password":"password","data":"data"}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'if_match': 'if_match_example',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PUT',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.ApiManagement/service/{service_name}/certificates/{certificate_id}'.format(resource_group_name='resource_group_name_example', service_name='service_name_example', certificate_id='certificate_id_example', subscription_id='subscription_id_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_certificates_delete(client):
    """Test case for certificates_delete

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'if_match': 'if_match_example',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.ApiManagement/service/{service_name}/certificates/{certificate_id}'.format(resource_group_name='resource_group_name_example', service_name='service_name_example', certificate_id='certificate_id_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_certificates_get(client):
    """Test case for certificates_get

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.ApiManagement/service/{service_name}/certificates/{certificate_id}'.format(resource_group_name='resource_group_name_example', service_name='service_name_example', certificate_id='certificate_id_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_certificates_list_by_service(client):
    """Test case for certificates_list_by_service

    
    """
    params = [('$filter', 'filter_example'),
                    ('$top', 56),
                    ('$skip', 56),
                    ('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.ApiManagement/service/{service_name}/certificates'.format(resource_group_name='resource_group_name_example', service_name='service_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

